requirements.txt
# requirements.txt
"""
openai>=1.3.0
replicate>=0.15.0
bark>=1.0.0
moviepy>=1.0.3
python-dotenv>=1.0.0
Pillow>=10.0.0
scipy>=1.11.0
numpy>=1.24.0
requests>=2.31.0
"""